feat: modify function for creating keyboard
